Proposal completed
