+++ sample/src/CustomLayoutDataGrid.tsx
@@ -10,7 +10,7 @@ import Typography from '@material-ui/core/Typography';
 import CheckBox from '@material-ui/icons/CheckBox';
 import CheckBoxOutlineBlank from '@material-ui/icons/CheckBoxOutlineBlank';
 import * as React from 'react';
-import { ColumnModel, formatDate } from 'tubular-common';
+import { ColumnModel, parseDateColumnValue } from 'tubular-common';
 import { FixedLinearProgress } from 'uno-material-ui/dist/FixedLinearProgress';
 import { useResolutionSwitch } from 'uno-react/lib/hooks/useResolutionSwitch';
 import { TbRowProps } from '../../src/BareBones/TbRow';
@@ -25,7 +25,7 @@ const CustomTbRow: React.FunctionComponent<TbRowProps> = ({ row, onRowClick }: T
     <TableRow hover={true} key={row.OrderID} onClick={onRowClick}>
         <TableCell padding="default">{row.OrderID} </TableCell>
         <TableCell padding="default">{row.CustomerName} </TableCell>
-        <TableCell padding="default">{formatDate(row.ShippedDate, 'M/d/yyyy h:mm a')}</TableCell>
+        <TableCell padding="default">{row.ShippedDate}</TableCell>
         <TableCell padding="default">{row.ShipperCity}</TableCell>
         <TableCell padding="default" align={'right'}>
             {row.Amount || 0}
