+++ src/BareBones/DetailComponentProps.ts
@@ -1,0 +1,3 @@
+export default interface DetailComponentProps {
+    row?: any;
+}
