+++ src/BareBones/index.ts
@@ -1,2 +1,3 @@
 export * from './TbMobileRow';
 export * from './TbRow';
+export * from './DetailComponentProps';
