--- src/DataGrid/DetailBaseComponent.tsx
@@ -1,6 +1,0 @@
-import * as React from 'react';
-import DetailComponent from '../DataGridInterfaces/DetailComponent';
-
-export type DetailBaseComponent = React.FunctionComponent<DetailComponent>;
-
-export default DetailBaseComponent;
