+++ src/DataGrid/index.ts
@@ -1,4 +1,3 @@
 export * from './DataGridTable';
 export * from './DataGridCard';
 export * from './DataGrid';
-export * from './DetailBaseComponent';
