--- src/DataGridInterfaces/DetailComponent.ts
@@ -1,3 +1,0 @@
-export default interface DetailComponent {
-    row?: any;
-}
